requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=/Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.txt /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
#
aiofiles==22.1.0
    # via -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
aiologger==0.7.0
    # via -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
fastapi==0.88.0
    # via -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.1
    # via -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
pydantic==1.10.2
    # via fastapi
python-dotenv==0.21.0
    # via
    #   -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
    #   uvicorn
pyyaml==6.0
    # via
    #   -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
    #   uvicorn
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.22.0
    # via fastapi
typing-extensions==4.4.0
    # via pydantic
uvicorn[standard]==0.20.0
    # via -r /Users/scanner/src/postmark-as-mail-tests/inbound_api/requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.18.1
    # via uvicorn
websockets==10.4
    # via uvicorn
